requirements.txt
python-telegram-bot==22.6
pandas==2.2.3
numpy==2.1.3
matplotlib==3.9.2
yfinance==0.2.43
SQLAlchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
requests==2.32.3